Refactored LongAdder to ConcurrentCounter
